refactor: Move all source code into 'src/'
The index.ts now lives among the rest of the source code in 'src/',
instead of weirdly being placed among tool configs.

This patch also contains additional modifications such as removal of the
second .eslintrc.yml in 'test/' in favor of an override, updating the
year range in LICENSE, changing the globs in TSConfig from 'exclude' to
'include', and ensuring whitespace in README.md follows the lint style.